refactor(P2.1): move VLP expr-rewriting group to render_plan/vlp_rewrite.rs
First Phase-2 §5.1 module MOVE (REFACTORING_SAFETY_PLAN.md §5.1, the
lowest-risk pure-transform group). Extracts the VLP expression-rewriting
cluster out of the 17.7K-line plan_builder_utils.rs god-file into a new
render_plan/vlp_rewrite.rs (796 lines).

Moved verbatim (ZERO logic edits):
- rewrite_vlp_aggregate_aliases (+ its private helper rewrite_expr_for_vlp_end_nodes)
- rewrite_render_expr_for_vlp_with_from_alias
- translate_db_columns_to_cypher_properties (fn -> pub(crate) fn: its sole
  caller stayed behind; visibility fix only)
- rewrite_render_expr_for_vlp_with_endpoint_info
- extract_vlp_alias_mappings

The old path pub(crate)-re-exports all five so the narrow inbound API
(plan_builder.rs + internal callers) is unaffected. No new import cycle:
the new module depends only on render_expr/errors/join_context/
graph_catalog, never back on plan_builder::RenderPlanBuilder.

Deliberately NOT done this slice (filed as follow-ups): D3 VLP expr-rewriter
generation dedup — pure move only, the file churned past the §5.1 line
numbers so a mechanical cut-paste is the safe choice.

Acceptance: 211 goldens + 1,082-query corpus sweep byte-identical; full
cargo test green; fmt + clippy --all-targets clean. Ratchet baseline updated
for the pure relocation of 4 is_denormalized structural-derivation tokens
(plan_builder_utils 9->5, vlp_rewrite 0->4, net zero — no new branching).

Docs: render_plan/AGENTS.md module map, REFACTORING_SAFETY_PLAN.md §9,
PRIORITIES.md §2/§4 updated in the same PR.
