+++ src/render_plan/mod.rs
@@ -20,6 +20,7 @@ pub mod types;
 pub mod utils;
 pub mod variable_scope;
 mod view_table_ref;
+pub(crate) mod vlp_rewrite;
 
 use errors::RenderBuildError;
 use render_expr::{ColumnAlias, OperatorApplication, RenderExpr};
